+++ src/app/core/model/admin/approval-or-rejection.model.ts
@@ -1,0 +1,6 @@
+export interface ApprovalOrRejectionConfirmation {
+  workplaceName: string;
+  approvalName: string;
+  approvalType: string;
+  isApproval: boolean;
+}
